backend/requirements.txt
Flask
SQLAlchemy
Flask-SQLAlchemy
Flask-Migrate
PyMySQL
Flask-Bcrypt
PyJWT
python-dotenv
flask-cors
cryptography
